requirements.txt
Django==2.2.14
djangorestframework==3.9.4
pytest==5.3.5
pytest-django==3.8.0
